% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{.dsj_option_safe}
\alias{.dsj_option_safe}
\title{Read dsjobs option safely (dsRadiomics doesn't import dsJobs)}
\usage{
.dsj_option_safe(name)
}
\description{
Read dsjobs option safely (dsRadiomics doesn't import dsJobs)
}
\keyword{internal}
