% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_hooks.R
\name{.radiomics_publisher}
\alias{.radiomics_publisher}
\title{Radiomics asset publisher (dsJobs plugin) -- collection-level}
\usage{
.radiomics_publisher(job_id, step, output_dir, db)
}
\description{
Radiomics asset publisher (dsJobs plugin) -- collection-level
}
\keyword{internal}
