% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrator.R
\name{.resolve_image_root}
\alias{.resolve_image_root}
\title{Resolve image root URI from manifest}
\usage{
.resolve_image_root(dataset_id)
}
\description{
Resolve image root URI from manifest
}
\keyword{internal}
