% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrator.R
\name{.resolve_mask_root}
\alias{.resolve_mask_root}
\title{Resolve mask root URI when using existing masks}
\usage{
.resolve_mask_root(dataset_id, segmenter)
}
\description{
Resolve mask root URI when using existing masks
}
\keyword{internal}
