% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{radiomicsCapabilitiesDS}
\alias{radiomicsCapabilitiesDS}
\title{Get Radiomics Capabilities}
\usage{
radiomicsCapabilitiesDS()
}
\value{
Named list of capabilities.
}
\description{
DataSHIELD AGGREGATE method.
}
