% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{radiomicsListModelsDS}
\alias{radiomicsListModelsDS}
\title{List Installed Models (no admin needed)}
\usage{
radiomicsListModelsDS()
}
\value{
Data.frame of installed models.
}
\description{
DataSHIELD AGGREGATE method.
}
