% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{register_segmentation_model}
\alias{register_segmentation_model}
\title{Register a segmentation model (admin utility)}
\usage{
register_segmentation_model(
  name,
  provider,
  task,
  path,
  python_deps = NULL,
  extra = list()
)
}
\description{
Register a segmentation model (admin utility)
}
